tuesday work

+++ anime-api/controllers/animeController.js
@@ -21,6 +21,11 @@ export let createAnime = async (request, response, next) => {
 
 export const getSingleAnime = async (request, response, next) => {
   const id = request.params.id;
+  if (!id) {
+    return response
+      .status(404)
+      .send({ message: "Anime cannot be found check again!" });
+  }
   try {
     const singleAnime = await Anime.findById(id);
     return response.status(200).json(singleAnime);
@@ -31,6 +36,11 @@ export const getSingleAnime = async (request, response, next) => {
 
 export const deleteSingleAnime = async (request, response, next) => {
   const id = request.params.id;
+  if (!id) {
+    return response
+      .status(404)
+      .send({ message: "Anime cannot be found check again!" });
+  }
   try {
     const deleteAnime = await Anime.findByIdAndDelete(id);
     return response.status(200).json(deleteAnime);
@@ -41,6 +51,11 @@ export const deleteSingleAnime = async (request, response, next) => {
 
 export const updateSingleAnime = async (request, response, next) => {
   const id = request.params.id;
+  if (!id) {
+    return response
+      .status(404)
+      .send({ message: "Anime cannot be found check again!" });
+  }
   try {
     const updateAnime = await Anime.findByIdAndUpdate(id, request.body, {
       new: true
